refactor: Controller 인터페이스 및 추상 클래스 패키지 위치 변경
